requirements.txt
streamlit==1.38.0
psycopg2-binary==2.9.9
pandas==2.2.2
openpyxl==3.1.5
bcrypt==4.1.3
python-dotenv==1.0.1
